Prioritize Debian below other sources
We want packages from custom repositories to be preferred over packages
in Debian proper, allowing us to customize packages that are in Debian
without maintaining a larger version number.

Note the same preferences may be needed during network-based software
updates on the device, although pointing directly at a Debian repository
from a device is not recommended.